% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_missing_values}
\alias{check_missing_values}
\title{Check for missing values in an expression matrix}
\usage{
check_missing_values(x)
}
\arguments{
\item{x}{A numeric matrix or data frame.}
}
\value{
\code{TRUE} if any missing values are present, \code{FALSE} otherwise.
}
\description{
Check for missing values in an expression matrix
}
\keyword{internal}
