% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyHNSC_interface.R
\name{classifyHNSC_interface}
\alias{classifyHNSC_interface}
\title{The Shiny interface of head and neck squamous cell carcinoma (HNSC) molecular subtypes prediction}
\usage{
classifyHNSC_interface()
}
\description{
A shiny interface for molecular subtype prediction in HNSCclassifier
}
\examples{
classifyHNSC_interface()
}
