% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HNSCclassifier-package.R
\docType{package}
\name{HNSCclassifier-package}
\alias{HNSCclassifier}
\alias{HNSCclassifier-package}
\title{HNSCclassifier: HNSCclassifier: An R Package to Predict Molecular Subtypes of Head and Neck Squamous Cell Carcinoma}
\description{
Description: This package provides a robust classifier to assign head and neck squamous cell carcinoma (HNSCC) samples to the established TCGA molecular subtypes (Atypical, Basal, Classical, Mesenchymal). It implements a pathway-level normalization strategy based on single-sample gene set enrichment (ssGSEA) and a pre-trained random forest model. The package includes a command-line function for batch prediction and an interactive Shiny web interface for users without programming experience.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ronlee12355/HNSCclassifier}
  \item Report bugs at \url{https://github.com/Ronlee12355/HNSCclassifier/issues}
}

}
\author{
\strong{Maintainer}: Jiang Li \email{ronlee12355@outlook.com}

}
\keyword{internal}
